sparrow: insar display improvements
- color handling and general setup
- float, int -> float64, int64 (numpy v1.24)
- right display of NaN values
- adopt CPTHandler for Kite Scene display purposes
- update internal functions

+++ src/gui/sparrow/elements/spheroid.py
@@ -113,7 +113,8 @@ class SpheroidElement(Element):
 
         if self._mesh:
             vertices = self._vertices0 * s[num.newaxis, :]
-            bc = num.matrix([[1, 0, 0], [0, 0, 1], [0, 1, 0]], dtype=num.float)
+            bc = num.matrix(
+                [[1, 0, 0], [0, 0, 1], [0, 1, 0]], dtype=num.float64)
             rot = num.dot(
                 bc,
                 num.dot(
